Made ParticleLayer inherit Atmosphere
