+++ docs/generate_rst_api.py
@@ -13,9 +13,9 @@ console = Console()
 
 # List of (module, variable) pairs
 FACTORIES = [
-    ("eradiate.radprops.particle_dist", "particle_distribution_factory"),
     ("eradiate.radprops.rad_profile", "rad_profile_factory"),
     ("eradiate.scenes.atmosphere", "atmosphere_factory"),
+    ("eradiate.scenes.atmosphere._particle_dist", "particle_distribution_factory"),
     ("eradiate.scenes.biosphere", "biosphere_factory"),
     ("eradiate.scenes.illumination", "illumination_factory"),
     ("eradiate.scenes.integrators", "integrator_factory"),
