+++ eradiate/radprops/__init__.py
@@ -3,7 +3,6 @@ Atmospheric radiative properties calculation package.
 """
 
 
-from .particle_dist import ParticleDistribution, particle_distribution_factory
 from .rad_profile import (
     AFGL1986RadProfile,
     ArrayRadProfile,
@@ -14,8 +13,6 @@ from .rad_profile import (
 __all__ = [
     "AFGL1986RadProfile",
     "ArrayRadProfile",
-    "ParticleDistribution",
     "US76ApproxRadProfile",
-    "particle_distribution_factory",
     "rad_profile_factory",
 ]
