+++ eradiate/scenes/atmosphere/__init__.py
@@ -1,10 +1,15 @@
 from ._core import Atmosphere, atmosphere_factory
 from ._heterogeneous import HeterogeneousAtmosphere
 from ._homogeneous import HomogeneousAtmosphere
+from ._particle_dist import ParticleDistribution, particle_distribution_factory
+from ._particles import ParticleLayer
 
 __all__ = [
     "Atmosphere",
     "HeterogeneousAtmosphere",
     "HomogeneousAtmosphere",
     "atmosphere_factory",
+    "particle_distribution_factory",
+    "ParticleDistribution",
+    "ParticleLayer",
 ]
